Implemented Collision Detection
Messed around with collision detection. Added a list of used tiles to
the map that is updated when tiles are populated. This is then checked
against to see if a meaningful tile is collided with.

For now, the file prints out which tile was landed on and makes sure to
remove said tile (using .pop and .index) from the list of used tiles.