Added the intro and added encounter reshuffling
We now have a separate intro.py file to completely handle the
introduction to the game. The introductory story part will be added
later.

Also set up the encounter section so that if you need more events than
we currently the have, the list of unusedEvents will be replenished
avoiding an errors that would occur with running out of events while
trying to populate maps.